refactor(commands/accounts): move commands to their own file
