+++ src-tauri/src/commands/accounts/mod.rs
@@ -1,0 +1,19 @@
+mod create_identity;
+mod get_accounts;
+mod has_nostr_wallet_connect_uri;
+mod login;
+mod logout;
+mod remove_nostr_wallet_connect_uri;
+mod set_active_account;
+mod set_nostr_wallet_connect_uri;
+mod update_account_onboarding;
+
+pub use create_identity::create_identity;
+pub use get_accounts::get_accounts;
+pub use has_nostr_wallet_connect_uri::has_nostr_wallet_connect_uri;
+pub use login::login;
+pub use logout::logout;
+pub use remove_nostr_wallet_connect_uri::remove_nostr_wallet_connect_uri;
+pub use set_active_account::set_active_account;
+pub use set_nostr_wallet_connect_uri::set_nostr_wallet_connect_uri;
+pub use update_account_onboarding::update_account_onboarding;
